dynamic deleting and adding routers

+++ src/main/java/socs/network/message/LinkDescription.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class LinkDescription implements Serializable {
     public String linkID;
     public int portNum;
-    public int tosMetrics;
+    public double tosMetrics;
 
     public String toString() {
         return linkID + "," + portNum + "," + tosMetrics;
